drm/i915: prevent the vgacon from ever reloading
It only leads to bloodshed and tears - we don't bother to restore a
working legacy vga hw setup.

On haswell with the new dynamic power well code this leads to even
more hilarity since for some configurations the hardware is simply no
longer there.

The actual implementation is a bit a hack - we realy on fbcon to kick
out the vgacon. To make this also work with I915_FBDEV=n (or FBCON=n)
and VGA_CONSOLE=y i915 already unregisters the vga console manually
early in the driver load sequence.

Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Paulo Zanoni <przanoni@gmail.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=67813
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>