HACK: iommu/arm-smmu: make a mismatch in CBs and CB interrupts non-fatal
When running under a hypervisor, the number of context banks visible to
Linux could be different from what we see when running standalone.  This
is problematic since the context bank interrupts are hard-coded into the
DT configuration, and we'd like to share that configuration whether
we're running under a hypervisor or not.  Remove a warning when a
mismatch in the number of CBs and CB interrupts is detected and don't
return failure while probing.

Note from Vivek Gautam about this patch:
  For the LC bootchain we now don't need this patch as well.  On the
  LA bootchain, one of the context bank is marked secure, so out of 64
  CBs HLOS was seeing only 63, and we needed this patch.

  With LC bootchain however, we are good. The kernel sees 64 CBs, and
  therefore we don't have a mismatch between the
  smmu->num_context_banks and smmu->num_context_irqs.

Change-Id: I55296ec3474db7aab8299f93736aae22bc210bd1
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>
Signed-off-by: Patrick Daly <pdaly@codeaurora.org>
Signed-off-by: Vivek Gautam <vivek.gautam@codeaurora.org>
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>
[dianders added HACK prefix and note from Vivek]
Signed-off-by: Douglas Anderson <dianders@chromium.org>