WIP: drm/msm/rd: allow adding addition msg to top of dump
For faults or hangs, it is nice to be able to include a bit more
information.

TODO: maybe rd_printf() style interface would make more sense.  It would
be nice to (for example) and the info from a5xx err irq handlers.  But
slightly complicated because we can't write into circbuf from IRQ as it
may need to block on userspace reading out the other side.