UPSTREAM: power: reset: msm: Clarify restart and poweroff
When PSHOLD in a Qualcomm platform is deasserted the PMIC will perform
either a power off or a restart of the system. The action to take is
configured in the PON block, which is controlled by a separate driver.

As the configuration logic was added to the pm8941-pwrkey driver the
comment in do_msm_poweroff() is no longer valid and the name
do_msm_restart() is misleading. Update the naming and drop the comment.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.co.uk>
(cherry picked from commit fd8b8f17d820328c0a2a46653e44e12a8cd450ba)

BUG=b:110849289
TEST=Pat your head and rub your tummy

Change-Id: Ia1875d0086bf30d9bb028d16df104932ba58972d
Signed-off-by: Douglas Anderson <dianders@chromium.org>