DO-NOT-SUBMIT: Hack it up so we can reboot cheza from the kernel
Right now "reboot" doesn't work on the command line.  That's because
it's not implemented properly in PSCI.  It might take a while before
we can get this done, so for now people who want reboot can pick this
patch into their local tree.

Q: Why can't this land?
A: Officially it's required to support reboot through PSCI.

Q: Why will it take a while for this to work in PSCI?
A: We need a GPIO to reboot, which means board-specific data in ATF.
   We've done it before, but it might be a while to get it all
   plumbed in.

BUG=chromium:795946
TEST=reboot

Change-Id: Ibabf2b46d495836253646b306422696acbc41a49
Signed-off-by: Douglas Anderson <dianders@chromium.org>