WIP: drm/msm: dsi: Handle dsi_bind() failures better
In dsi_bind() if msm_dsi_has_valid_device() returned false then we
would orphan the msm_dsi device (dsi_unbind() would never be able to
find it, for instace).  That seems pretty bad.

I tried call dsi_destroy() and returning an error, but unfortunately
this made the display fail to come up on cheza (huh???) so it appears
that something was relying on the behavior.

For now at least print an error message and leave the commented out
dsi_destroy() / error to make it obvious that something seems wrong.

NOTE: originally I came across this case by trying to make sense of
meaningless crashes when I had a bad device tree.

Change-Id: I6a929c4784275b2a7564f5238f6a4b5b1d806651
Signed-off-by: Douglas Anderson <dianders@chromium.org>