HACK: Drop DFS usage from spi driver
We don't have DFS support in the clk driver yet, so we can't assume that
the frequency table the clk driver tells us about is actually programmed
into the DFS tables of SPI. Drop the logic and derive all our clk rates
from 19.2 MHz right now. Later on when DFS lands in GCC we can get rid
of this patch.

BUG=None
TEST=Build and boot

Change-Id: I57feeccc0a3bca508c12daea7e9c3d4da61b7447
Signed-off-by: Stephen Boyd <swboyd@chromium.org>