FROMLIST: soc: qcom: geni: Make version macros simpler
This macro doesn't work, because it hides a local variable inside of the
macro to hold the version and that variable name is called 'ver' and
'version' sometimes.

Let's change this to be more explicit. Introduce three macros for the
major, minor, and step of the version, and require callers to pass the
version in to get the part of the version out. This way we don't hide
local variables inside macros and things are less evil overall.

BUG=None
TEST=Build and boot

Cc: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
Cc: Sagar Dharia <sdharia@codeaurora.org>
Cc: Girish Mahadevan <girishm@codeaurora.org>
(am from https://patchwork.kernel.org/patch/10412417/)
Change-Id: I65e7707da193b4968d233340b0fbed1919fe631f
Signed-off-by: Stephen Boyd <swboyd@chromium.org>