FROMLIST: drm/msm/adreno: Load the firmware before bringing up the hardware
Failure to load firwmare is the primary reason to fail adreno_load_gpu().
Try to load it first before going into the hardware initialization code and
unwinding it. This is important for a6xx because the GMU gets loaded from
the runtime power code and it is more costly to fail in that path because
of missing firmware.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>

BUG=72483556
TEST=Build and boot

Change-Id: I2eb8377a4998e84d0f4845d35f8a7929c7a42ee6
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
(am from https://patchwork.kernel.org/patch/10458653/)