FROMLIST: drm/msm: Add A6XX device support
Add support for the A6XX family of Adreno GPUs. The biggest addition
is the GMU (Graphics Management Unit) which takes over most of the
power management of the GPU itself but in a ironic twist of fate
needs a goodly amount of management itself. Add support for the
A6XX core code, the GMU and the HFI (hardware firmware interface)
queue that the CPU uses to communicate with the GMU.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>

BUG=72483556
TEST=Build and boot

Change-Id: I233a05a29c9dea7b2112abc134ee31ee0f2bd4e3
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
(am from https://patchwork.kernel.org/patch/10458663/)