FROMLIST: drm/msm: Add a helper function to parse clock names
Add a helper function to parse the clock names and set up
the bulk data so we can take advantage of the bulk clock
functions instead of rolling our own. This is added
as a helper function so the upcoming a6xx GMU code can
also take advantage of it.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>

BUG=72483556
TEST=Build and boot

Change-Id: Ie61ef053162847057f0eb0185c02e8fc4ff3b74d
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
(am from https://patchwork.kernel.org/patch/10458657/)