CHROMIUM: config: Turn on CONFIG_POWER_RESET_MSM_DOWNLOAD_MODE
This will be used for crash dumps.  This is the equivalnet of:

  echo 'CONFIG_POWER_RESET_MSM=y' | tee -a \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config \
    chromeos/config/arm64/chromiumos-arm64.flavour.config
  echo 'CONFIG_POWER_RESET_MSM_DOWNLOAD_MODE=y' | tee -a \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config \
    chromeos/config/arm64/chromiumos-arm64.flavour.config
  ./chromeos/scripts/kernelconfig olddefconfig

...atop a config that was normalized.

NOTE: Much of the CONFIG_POWER_RESET_MSM looks like it's not what we
want since most of the driver is effectively a no-op when PSCI is
enabled (restart handlers aren't called when PSCI is used.  The new
part (CONFIG_POWER_RESET_MSM_DOWNLOAD_MODE) might do something because
that seems to happen at probe time, but the rest of the driver
shouldn't won't have any effect.

...well, to be honest, we have a hack in our tree ("DO-NOT-SUBMIT:
Hack it up so we can reboot cheza from the kernel") that will make
restart handlers be called, but that same hack registers a higher
priority handler anyway.

BUG=b:110849289
TEST=Pat your head and rub your tummy

Change-Id: I628657404fc973d7a8185ee17eb5f56aa62911aa
Signed-off-by: Douglas Anderson <dianders@chromium.org>