BACKPORT: FROMLIST: interconnect: Add generic on-chip interconnect API
This patch introduces a new API to get requirements and configure the
interconnect buses across the entire chipset to fit with the current
demand.

The API is using a consumer/provider-based model, where the providers are
the interconnect buses and the consumers could be various drivers.
The consumers request interconnect resources (path) between endpoints and
set the desired constraints on this data flow path. The providers receive
requests from consumers and aggregate these requests for all master-slave
pairs on that path. Then the providers configure each participating in the
topology node according to the requested data flow path, physical links and
constraints. The topology could be complicated and multi-tiered and is SoC
specific.

Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>
Reviewed-by: Evan Green <evgreen@chromium.org>
(am from https://patchwork.kernel.org/patch/10562799/)

Makefile conflicts occurred due to nearby unrelated changes.
Conflicts:
   drivers/Makefile

BUG=b:80591049
TEST=Build kernel for cheza

Change-Id: Idd574aa03e7683de41e04b5a5b3f2c48a3e58d2b
Signed-off-by: Evan Green <evgreen@chromium.org>

Conflicts:
  drivers/Kconfig
  drivers/Makefile
  [rebase419, mka@: context deltas]

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>