FROMLIST: drm/msm/dpu: Integrate interconnect API in MDSS
The interconnect framework is designed to provide a
standard kernel interface to control the settings of
the interconnects on a SoC.

The interconnect API uses a consumer/provider-based model,
where the providers are the interconnect buses and the
consumers could be various drivers.

MDSS is one of the interconnect consumers which uses the
interconnect APIs to get the path between endpoints and
set its bandwidth/latency/QoS requirements for the given
interconnected path.

Changes in v2:
	- Remove error log and unnecessary check (Jordan Crouse)

Signed-off-by: Sravanthi Kollukuduru <skolluku@codeaurora.org>
(am from https://patchwork.kernel.org/patch/10634271/)
(also found at https://lkml.kernel.org/r/20181010092434.853-3-skolluku@codeaurora.org)

BUG=b:117981951, b:80591049
TEST=Display comes up

Change-Id: Icf22d24191fe7fb3aa410a36b8967ff1676fd0f6
Signed-off-by: Douglas Anderson <dianders@chromium.org>