WIP: drm/msm: add WFI before capturing counters
Otherwise the CP can execute the CP_REG_TO_MEM in parallel with the GPU
executing the previous draw/resolve/blit/etc.

But adding a pair of WFI in the submit path isn't ideal for performance,
so make all submit_retired stats tracking, including the REG_TO_MEM and
WFIs, conditional on the trace being enabled.

This does have the upshot that the first few traces after enabling will
not have snapshot the begin/end counter values, and so will have bogus
elapsed duration, etc.  Which could be avoided by tracking that the
start/end counters are valid, but I'm not sure it is worth the effort.