WIP: iommu: nerf automatic DMA domain attach
This "feature" plays badly with devices enabled in bootloader, such as
display device with active scanout.  Attaching domain while dma is
already active causes iommu faults.  Since we can't put the toothpaste
back in the tube on this mis-feature, just blacklist devices where this
is known to be a problem.